backend/app/requirements.txt
Flask==2.0.1
docker==6.0.0
jsonpickle==2.2.0
Flask-PyMongo==2.3.0
pydantic==1.8.1
Flask-Login==0.6.2
Flask-Cors==3.0.10
